+++ src/proto/active_packetizer.rs
@@ -1,12 +1,12 @@
 use super::{request, watch::WatchType, Request, Response};
 use crate::{WatchedEvent, WatchedEventType, ZkError};
 use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
-use failure::format_err;
 use futures::{
     channel::{mpsc, oneshot},
     ready,
 };
 use pin_project::pin_project;
+use snafu::{Snafu, Whatever};
 use std::collections::HashMap;
 use std::{
     future::Future,
@@ -17,6 +17,27 @@ use std::{
 };
 use tokio::io::{AsyncRead, AsyncWrite, ReadBuf};
 use tracing::{debug, info, instrument, trace};
+
+#[derive(Debug, Snafu)]
+pub enum Error {
+    #[snafu(transparent)]
+    Io { source: std::io::Error },
+    #[snafu(transparent)]
+    Whatever { source: Whatever },
+
+    #[snafu(display("connection closed with {len} bytes left in buffer: {buf:x?}", len = buf.len()))]
+    ConnectionClosed { buf: Vec<u8> },
+
+    #[snafu(display("Not exiting, but server closed connection"))]
+    ServerClosedConnection,
+    #[snafu(display("outstanding requests, but response channel closed"))]
+    ResponseChannelClosedPrematurely,
+
+    #[snafu(display("bad response to ping: {error_code:?}"))]
+    BadPing { error_code: ZkError },
+    #[snafu(display("failed to close session: {error_code:?}"))]
+    CloseSession { error_code: ZkError },
+}
 
 #[pin_project]
 pub(super) struct ActivePacketizer<S> {
@@ -56,6 +77,8 @@ pub(super) struct ActivePacketizer<S> {
     pub(super) session_id: i64,
     pub(super) password: Vec<u8>,
 }
+
+type ReplySender = oneshot::Sender<Result<Response, ZkError>>;
 
 impl<S> ActivePacketizer<S>
 where
@@ -104,7 +127,7 @@ where
     #[allow(clippy::type_complexity)]
     fn enqueue_impl(
         outbox: &mut Vec<u8>,
-        reply: &mut HashMap<i32, (request::OpCode, oneshot::Sender<Result<Response, ZkError>>)>,
+        reply: &mut HashMap<i32, (request::OpCode, ReplySender)>,
         xid: i32,
         item: Request,
         tx: oneshot::Sender<Result<Response, ZkError>>,
@@ -164,7 +187,7 @@ where
         mut self: Pin<&mut Self>,
         cx: &mut Context,
         exiting: bool,
-    ) -> Poll<Result<(), failure::Error>>
+    ) -> Poll<Result<(), Error>>
     where
         S: AsyncWrite,
     {
@@ -191,11 +214,7 @@ where
                 .reset(tokio::time::Instant::now() + *this.timeout);
         }
 
-        ready!(this
-            .stream
-            .as_mut()
-            .poll_flush(cx)
-            .map_err(failure::Error::from)?);
+        ready!(this.stream.as_mut().poll_flush(cx)?);
 
         if exiting {
             debug!("shutting down writer");
@@ -210,7 +229,7 @@ where
         mut self: Pin<&mut Self>,
         cx: &mut Context,
         default_watcher: &mut mpsc::UnboundedSender<WatchedEvent>,
-    ) -> Poll<Result<(), failure::Error>>
+    ) -> Poll<Result<(), Error>>
     where
         S: AsyncRead,
     {
@@ -234,11 +253,12 @@ where
                         this.inbox.truncate(read_from + n);
                         if n == 0 {
                             if self.inlen() != 0 {
-                                return Poll::Ready(Err(format_err!(
-                                    "connection closed with {} bytes left in buffer: {:x?}",
-                                    self.inlen(),
-                                    &self.inbox[self.instart..]
-                                )));
+                                return Poll::Ready(
+                                    ConnectionClosedSnafu {
+                                        buf: &self.inbox[self.instart..],
+                                    }
+                                    .fail(),
+                                );
                             } else {
                                 // Server closed session with no bytes left in buffer
                                 debug!("server closed connection");
@@ -292,7 +312,7 @@ where
                     // XXX: in theory, server should now shut down receive end
                     trace!("got response to CloseSession");
                     if let Some(e) = err {
-                        return Poll::Ready(Err(format_err!("failed to close session: {:?}", e)));
+                        return Poll::Ready(CloseSessionSnafu { error_code: e }.fail());
                     }
                 } else if xid == -1 {
                     // watch event
@@ -347,7 +367,7 @@ where
                     // response to ping -- empty response
                     trace!("got response to heartbeat");
                     if let Some(e) = err {
-                        return Poll::Ready(Err(format_err!("bad response to ping: {:?}", e)));
+                        return Poll::Ready(BadPingSnafu { error_code: e }.fail());
                     }
                 } else {
                     // response to user request
@@ -425,7 +445,7 @@ where
         cx: &mut Context,
         exiting: bool,
         default_watcher: &mut mpsc::UnboundedSender<WatchedEvent>,
-    ) -> Poll<Result<(), failure::Error>> {
+    ) -> Poll<Result<(), Error>> {
         let r = self.as_mut().poll_read(cx, default_watcher)?;
 
         let mut this = self.as_mut().project();
@@ -461,12 +481,8 @@ where
                 debug!("packetizer done");
                 Poll::Ready(Ok(()))
             }
-            (Poll::Ready(()), Poll::Ready(())) => Poll::Ready(Err(format_err!(
-                "Not exiting, but server closed connection"
-            ))),
-            (Poll::Ready(()), _) => Poll::Ready(Err(format_err!(
-                "outstanding requests, but response channel closed"
-            ))),
+            (Poll::Ready(()), Poll::Ready(())) => Poll::Ready(ServerClosedConnectionSnafu.fail()),
+            (Poll::Ready(()), _) => Poll::Ready(ResponseChannelClosedPrematurelySnafu.fail()),
             _ => Poll::Pending,
         }
     }